bp,bs,ur,ear
need to deal with 9 in some cases. In the meantime, I am not very confident with the proper way to generate bp, bs, ur, ear variables. I generate c_anc_bp in several ways in this comment. I would appreciate it if you could let me know whether any of this is correct.